IE fix: attachments list
Created two different blocks of HTML to represent attachments list as a
table and as a list of thumbnails. This is done due to IE6 bug: it's
unable to render table/tr/td/th as block elements even if display: block
is set to them in styles. Actually display CSS property of
<table>/<tr>/<td>/<th> is immutable in IE, even if you change it in
styles, IE will still render these elements as table.